requirements.txt
pytesseract
opencv-python
Pillow
